working on GUI
